+++ r2-navigator-swift/EPUB/EPUBFixedSpreadView.swift
@@ -1,12 +1,7 @@
 //
-//  EPUBFixedSpreadView.swift
-//  r2-navigator-swift
-//
-//  Created by Mickaël Menu on 09.04.19.
-//
-//  Copyright 2019 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Copyright 2020 Readium Foundation. All rights reserved.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import Foundation
@@ -40,7 +35,7 @@ final class EPUBFixedSpreadView: EPUBSpreadView {
         
         // Loads the wrapper page into the web view.
         let spreadFile = "fxl-spread-\(spread.pageCount.rawValue)"
-        if let wrapperPageURL = Bundle(for: type(of: self)).url(forResource: spreadFile, withExtension: "html"), let wrapperPage = try? String(contentsOf: wrapperPageURL, encoding: .utf8) {
+        if let wrapperPageURL = Bundle.module.url(forResource: spreadFile, withExtension: "html"), let wrapperPage = try? String(contentsOf: wrapperPageURL, encoding: .utf8) {
             // The publication's base URL is used to make sure we can access the resources through the iframe with JavaScript.
             webView.loadHTMLString(wrapperPage, baseURL: publication.baseURL)
         }
@@ -62,6 +57,11 @@ final class EPUBFixedSpreadView: EPUBSpreadView {
             return
         }
         webView.evaluateJavaScript("spread.load(\(spread.jsonString(for: publication)));")
+    }
+
+    override func spreadDidLoad() {
+        super.spreadDidLoad()
+        goToCompletions.complete()
     }
 
     /// Layouts the resource to fit its content in the bounds.
@@ -100,8 +100,24 @@ final class EPUBFixedSpreadView: EPUBSpreadView {
             loadSpread()
         }
     }
-    
-    
+
+    /// MARK: - Location and progression
+
+    private let goToCompletions = CompletionList()
+
+    override func go(to location: PageLocation, completion: (() -> ())?) {
+        // Fixed layout resources are always fully visible so we don't use the location.
+        guard let completion = completion else {
+            return
+        }
+
+        if spreadLoaded {
+            DispatchQueue.main.async(execute: completion)
+        } else {
+            goToCompletions.add(completion)
+        }
+    }
+
     /// MARK: - Scripts
     
     private static let fixedScript = loadScript(named: "fixed")