+++ CHANGELOG.md
@@ -2,14 +2,44 @@
 
 All notable changes to this project will be documented in this file.
 
-**Warning:** Features marked as *experimental* may change or be removed in a future release without notice. Use with
+**Warning:** Features marked as *alpha* may change or be removed in a future release without notice. Use with
 *caution.
 
 ## [Unreleased]
 
-### Changed
+### Added
 
-* CocoaPods is not supported anymore.
+* The EPUB navigator is now able to navigate to a `Locator` using its `text` context. This is useful for search results or highlights missing precise locations.
+* New `EPUBNavigatorViewController.evaluateJavaScript()` API to run a JavaScript on the currently visible HTML resource.
+* Support for Swift Package Manager (contributed by [@stevenzeck](https://github.com/readium/r2-navigator-swift/pull/176)).
+
+### Deprecated
+
+* Removed `navigator(_:userContentController:didReceive:)` which is actually not needed since you can provide your own `WKScriptMessageHandler` to `WKUserContentController`.
+
+### Fixed
+
+* Fixed receiving `EPUBNavigatorDelegate.navigator(_:setupUserScripts:)` for the first web view.
+
+
+## [2.0.0]
+
+### Deprecated
+
+* All APIs deprecated in previous versions are now unavailable.
+
+
+## [2.0.0-beta.2]
+
+### Added
+
+* New `EPUBNavigatorDelegate` APIs to inject custom JavaScript.
+  * Override `navigator(_:setupUserScripts:)` to register additional user script to the `WKUserContentController` of each web view.
+  * Override `navigator(_:userContentController:didReceive:)` to receive callbacks from your scripts.
+
+### Fixed
+
+* Optimized performances of preloaded EPUB resources.
 
 
 ## [2.0.0-beta.1]
@@ -65,3 +95,5 @@ progression. Now if no reading progression is set, the `effectiveReadingProgress
 [2.0.0-alpha.1]: https://github.com/readium/r2-navigator-swift/compare/1.2.6...2.0.0-alpha.1
 [2.0.0-alpha.2]: https://github.com/readium/r2-navigator-swift/compare/2.0.0-alpha.1...2.0.0-alpha.2
 [2.0.0-beta.1]: https://github.com/readium/r2-navigator-swift/compare/2.0.0-alpha.2...2.0.0-beta.1
+[2.0.0-beta.2]: https://github.com/readium/r2-navigator-swift/compare/2.0.0-beta.1...2.0.0-beta.2
+[2.0.0]: https://github.com/readium/r2-navigator-swift/compare/2.0.0-beta.2...2.0.0